+++ AdventOfCode/Y23/D06/Race.cs
@@ -1,3 +1,3 @@
 ﻿namespace AdventOfCode.Y23.D06;
 
-public record Race(int Time, int RecordDistance);
+public record Race(long Time, long RecordDistance);
